requirements.txt
lark-parser
matrix-nio[e2e]
